quanlikhuyenMai

+++ src/components/admin/account/ViewAccount.js
@@ -8,7 +8,7 @@ function VỉewAccount() {
     const [viewAccount, setAccount] = useState([]);
     useEffect(() => {
         let isMounted = true;
-        document.title = "View Account";
+        document.title = "Danh sách tài khoản";
 
         axios.get(`/api/view-account`).then(res => {
             if (isMounted) {
@@ -43,42 +43,61 @@ function VỉewAccount() {
     }
     else {
         display_Accountdata = viewAccount.map((item, index) => {
+            if (item.role_as == 1) {
+                return (
+                    <tr id={item.id} key={index}>
+                        <td className='fs-4 text'>{item.id}</td>
+                        <td className='fs-4 text'>{item.name}</td>
+                        <td className='fs-4 text'>{item.email}</td>
+                        <td className='fs-4 text'>{item.address}</td>
+                        <td className='fs-4 text'>{item.phone}</td>
+                        <td className='fs-4 text' style={{color:'red',fontWeight:"bold",fontSize:"1.6rem"}} >Quản trị viên</td>
+                        <td></td>
+                    </tr >
+                )
+            }
+            else {
+                return (
+                    <tr id={item.id} key={index}>
+                        <td className='fs-4 text'>{item.id}</td>
+                        <td className='fs-4 text'>{item.name}</td>
+                        <td className='fs-4 text'>{item.email}</td>
+                        <td className='fs-4 text'>{item.address}</td>
+                        <td className='fs-4 text'>{item.phone}</td>
+                        {item.role_as == 2 ? 
+                        <td className='fs-4 text' style={{color:'blue',fontWeight:"bold",fontSize:"1.6rem"}}>Nhân viên</td> 
+                        : <td className='fs-4 text' style={{color:'#0ccf0f',fontWeight:"bold",fontSize:"1.6rem"}}>Khách hàng</td>}
+                        <td><button type="button" onClick={(e) => {
+                            swal({
+                                title: "Thông báo!",
+                                text: "Bạn có chắc muốn xóa không!",
+                                icon: "warning",
+                                buttons: [
+                                    'No',
+                                    'Yes'
+                                ],
+                                dangerMode: true,
+                            }).then(function (isConfirm) {
+                                if (isConfirm) {
+                                    deleteAccount(e, item.id);
+                                } else {
 
-            return (
-                <tr id={item.id} key={index}>
-                    <td className='fs-4 text'>{item.id}</td>
-                    <td className='fs-4 text'>{item.name}</td>
-                    <td className='fs-4 text'>{item.email}</td>
-                    <td className='fs-4 text'>{item.address}</td>
-                    <td className='fs-4 text'>{item.phone}</td>
-                    <td className='fs-4 text'>{item.role_as === 1 ? "quản trị viên" : "khách hàng"}</td>
-                    <td><button type="button" onClick={(e) => {
-                        swal({
-                            title: "Thông báo!",
-                            text: "Bạn có chắc muốn xóa không!",
-                            icon: "warning",
-                            buttons: [
-                                'No',
-                                'Yes'
-                            ],
-                            dangerMode: true,
-                        }).then(function (isConfirm) {
-                            if (isConfirm) {
-                                deleteAccount(e, item.id);
-                            } else {
-
-                            }
-                        })
-                    }} className='btn btn-danger btn-sm fs-4 text'>Xóa</button></td>
-                </tr >
-            )
+                                }
+                            })
+                        }} className='btn btn-danger btn-sm fs-4 text'>Xóa</button></td>
+                    </tr >
+                )
+            }
         });
     }
+
+
     return (
         <div className="container px-4 mt-3">
             <div className="card">
                 <div className="card-header">
                     <h2>Danh sách tài khoản
+                        <Link to="/admin/add-account" className="btn btn-primary btn-sm float-end  fs-4 text">Thêm tài khoản</Link>
                     </h2>
                 </div>
                 <div className="card-body">
@@ -90,8 +109,8 @@ function VỉewAccount() {
                                     <th>Tên </th>
                                     <th>Email</th>
                                     <th>Địa chỉ</th>
-                                    <th>số điện thoại</th>
-                                    <th>quản trị viên | khách hàng</th>
+                                    <th>Số điện thoại</th>
+                                    <th>Phân quyền</th>
                                     <th>Xóa </th>
                                 </tr>
                             </thead>
@@ -105,5 +124,6 @@ function VỉewAccount() {
         </div>
     );
 }
+
 
 export default VỉewAccount;
